backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
numpy==1.24.3
reportlab==4.0.4
Pillow==11.3.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
